feat: Add QuerierWithCtx interface to handle querier with context
This commit adds the QuerierWithCtx interface to the x/wasm/types package. The interface extends the wasmvm.Querier interface and includes a GetCtx() method to retrieve the context associated with the querier. This interface will be used to handle querier operations with context in the WasmEngine implementation.

This change is necessary to support the ongoing development of the x/wasm module and improve the functionality of the WasmEngine for smart contract execution.