feat: Add QuerierWithCtx, ContractGasRecorder, and ContractGasMeter types
This commit adds the QuerierWithCtx interface to handle querier operations with context in the x/wasm/types package. It also introduces two new types, ContractGasRecorder and ContractGasMeter, which will be used to handle gas consumption and metering for smart contracts. These changes are part of the ongoing development of the x/wasm module and aim to improve the functionality of the WasmEngine for smart contract execution.